agent/requirements.txt
requests==2.31.0
python-decouple==3.8
pyyaml==6.0.1

# Threat Detection
scapy==2.5.0
dpkt==1.9.7
pyshark==0.6.0
yara-python==4.3.0
pefile==2023.2.7

# Malware Analysis
python-magic==0.4.27
pycryptodome==3.19.0

# ML Models
scikit-learn==1.3.2
numpy==1.26.2
joblib==1.3.2

# System Monitoring
psutil==5.9.6
netifaces==0.11.0

# Logging
python-json-logger==2.0.7

# File Operations
pyaes==1.6.1
filetype==1.2.0

# Parsing & Analysis
lxml==4.9.3
beautifulsoup4==4.12.2
